refactor(Template): no need to externalize renderTemplate in utils
It belongs to the component right now, only used inside the component.
