+++ components/Template.js
@@ -1,7 +1,5 @@
 var React = require('react');
 
-var renderTemplate = require('../lib/utils').renderTemplate;
-
 class Template extends React.Component {
   render() {
     var content = renderTemplate({
@@ -67,6 +65,46 @@ function transformData(fn, templateKey, originalData) {
   }
 
   throw new Error('`transformData` must be a function or an object');
+}
+
+function renderTemplate({template, compileOptions, helpers, data}) {
+  var hogan = require('hogan.js');
+  var forOwn = require('lodash/object/forOwn');
+  var content;
+
+  if (typeof template !== 'string' && typeof template !== 'function') {
+    throw new Error('Template must be `string` or `function`');
+  }
+
+  if (typeof template === 'function') {
+    content = template(data);
+  }
+
+  if (typeof template === 'string') {
+    data = addTemplateHelpersToData(data);
+
+    content = hogan.compile(template, compileOptions).render(data);
+  }
+
+  // We add all our template helper methods to the template as lambdas. Note
+  // that lambdas in Mustache are supposed to accept a second argument of
+  // `render` to get the rendered value, not the literal `{{value}}`. But
+  // this is currently broken (see
+  // https://github.com/twitter/hogan.js/issues/222).
+  function addTemplateHelpersToData(templateData) {
+    templateData.helpers = {};
+    forOwn(helpers, (method, name) => {
+      templateData.helpers[name] = function() {
+        return (text) => {
+          var render = (value) => hogan.compile(value, compileOptions).render(this);
+          return method.call(this, text, render);
+        };
+      };
+    });
+    return data;
+  }
+
+  return content;
 }
 
 module.exports = Template;
